define the default value of status as empty string

+++ libpredweb/qd_fe_common.py
@@ -1130,6 +1130,7 @@ def GetResult(jobid, g_params):#{{{
             pass
         isSuccess = False
         isFinish_remote = False
+        status = ""
         if len(rtValue) >= 1:
             ss2 = rtValue[0]
             if len(ss2)>=3:
@@ -1289,7 +1290,7 @@ def GetResult(jobid, g_params):#{{{
             # running) delete it and try to resubmit it. This solved the
             # problem of dead jobs in the remote server due to server
             # rebooting)
-            if status != "Running" and time_in_remote_queue > g_params['MAX_TIME_IN_REMOTE_QUEUE']:
+            if status != "Running" and status != "" and time_in_remote_queue > g_params['MAX_TIME_IN_REMOTE_QUEUE']:
                 # delete the remote job on the remote server
                 try:
                     rtValue2 = myclient.service.deletejob(remote_jobid)
